requirements.txt
streamlit==1.22.0
folium==0.14.0
streamlit-folium==0.11.0
supabase==1.0.3
Pillow==9.5.0
google-generativeai==0.1.0
requests==2.28.2